Import QIIME2 Classification Table

Import taxonomy results exported from the QIIME2 Bayesian classifiers

import_qiime2_tax_table(in_file)

Arguments

in_file

A tab-delimited classification table output by QIIME2

Value

A phyloseq tax_table object

Details

This function expects up to 7 ranks (Domain, Phylum, Class, Order, Family, Genus and Species) but determines the number actually in the file.

Depending on how the QIIME2 classifier is built, the ranks in the Taxon field of the classification result may be separated by a semicolon or by a semicolon plus a space. This function correctly parses the Taxon field in both cases.

References

Bolyen E, Rideout JR, Dillon MR, Bokulich NA, et al. 2019. Reproducible, interactive, scalable and extensible microbiome data science using QIIME 2. Nat Biotechnol 37:852-857.

Examples

## Example with ranks in Taxon field separated by semicolon plus a space.
taxonomy_file <- system.file("extdata", "qiime2_table.tsv", package = "speedytax")
example_tax_table <- import_qiime2_tax_table(in_file = taxonomy_file)
example_tax_table
#> Taxonomy Table:     [10 taxa by 6 taxonomic ranks]:
#>                                  Domain       Phylum            
#> 7ff346973a282aa55de296afdb5d74af "d_Bacteria" "p_Actinomycetota"
#> 0f18144d308ada95632ab5193d92073f "d_Bacteria" "p_Pseudomonadota"
#> 496ecde24f9ab698992413d3d4f04b5f "d_Bacteria" "p_Pseudomonadota"
#> 9af3467db68cf6063627304cecd46a65 "d_Bacteria" "p_Pseudomonadota"
#> 07da22cab5c55a4badb462bcbbded642 "d_Bacteria" "p_Pseudomonadota"
#> 1bc599cac5647c5c30496e93ca551557 "d_Bacteria" "p_Pseudomonadota"
#> 08bdc6c81f6b9daab3d132b54e1d45bc "d_Bacteria" "p_Actinomycetota"
#> f84a5455a4f6fb3979f21796cdff4da4 "d_Bacteria" "p_Pseudomonadota"
#> c2a787cf867ebfdce216002bcc8a94e1 "d_Bacteria" "p_Pseudomonadota"
#> a0b1f105f827e856c5fdb5f98eaecc09 "d_Bacteria" "p_Pseudomonadota"
#>                                  Class                   Order              
#> 7ff346973a282aa55de296afdb5d74af "c_Actinomycetia"       "o_Actinomycetales"
#> 0f18144d308ada95632ab5193d92073f "c_Gammaproteobacteria" "o_Pseudomonadales"
#> 496ecde24f9ab698992413d3d4f04b5f "c_Gammaproteobacteria" "o_Xanthomonadales"
#> 9af3467db68cf6063627304cecd46a65 "c_Gammaproteobacteria" "o_Pseudomonadales"
#> 07da22cab5c55a4badb462bcbbded642 "c_Alphaproteobacteria" "o_Rhizobiales"    
#> 1bc599cac5647c5c30496e93ca551557 "c_Gammaproteobacteria" "o_Pseudomonadales"
#> 08bdc6c81f6b9daab3d132b54e1d45bc "c_Actinomycetia"       "o_Mycobacteriales"
#> f84a5455a4f6fb3979f21796cdff4da4 "c_Alphaproteobacteria" "o_Caulobacterales"
#> c2a787cf867ebfdce216002bcc8a94e1 "c_Alphaproteobacteria" "o_Rhizobiales"    
#> a0b1f105f827e856c5fdb5f98eaecc09 "c_Gammaproteobacteria" "o_Burkholderiales"
#>                                  Family                 Genus               
#> 7ff346973a282aa55de296afdb5d74af "f_Micrococcaceae"     "g_Arthrobacter"    
#> 0f18144d308ada95632ab5193d92073f "f_Pseudomonadaceae"   "g_Pseudomonas_E"   
#> 496ecde24f9ab698992413d3d4f04b5f "f_Xanthomonadaceae"   "g_Stenotrophomonas"
#> 9af3467db68cf6063627304cecd46a65 "f_Pseudomonadaceae"   "g_Pseudomonas_E"   
#> 07da22cab5c55a4badb462bcbbded642 "f_Rhizobiaceae"       "g_Phyllobacterium" 
#> 1bc599cac5647c5c30496e93ca551557 "f_Pseudomonadaceae"   "g_Pseudomonas_E"   
#> 08bdc6c81f6b9daab3d132b54e1d45bc "f_Mycobacteriaceae"   "g_Rhodococcus"     
#> f84a5455a4f6fb3979f21796cdff4da4 "f_Caulobacteraceae"   "g_Brevundimonas"   
#> c2a787cf867ebfdce216002bcc8a94e1 "f_Rhizobiaceae"       "g_Neorhizobium"    
#> a0b1f105f827e856c5fdb5f98eaecc09 "f_Burkholderiaceae_B" "g_Variovorax"      
## Example with ranks in Taxon field separated by semicolon only.
taxonomy_file <- system.file("extdata", "qiime2_table_a.tsv", package = "speedytax")
example_tax_table <- import_qiime2_tax_table(in_file = taxonomy_file)
example_tax_table
#> Taxonomy Table:     [10 taxa by 6 taxonomic ranks]:
#>                                  Domain       Phylum            
#> 7ff346973a282aa55de296afdb5d74af "d_Bacteria" "p_Actinomycetota"
#> 0f18144d308ada95632ab5193d92073f "d_Bacteria" "p_Pseudomonadota"
#> 496ecde24f9ab698992413d3d4f04b5f "d_Bacteria" "p_Pseudomonadota"
#> 9af3467db68cf6063627304cecd46a65 "d_Bacteria" "p_Pseudomonadota"
#> 07da22cab5c55a4badb462bcbbded642 "d_Bacteria" "p_Pseudomonadota"
#> 1bc599cac5647c5c30496e93ca551557 "d_Bacteria" "p_Pseudomonadota"
#> 08bdc6c81f6b9daab3d132b54e1d45bc "d_Bacteria" "p_Actinomycetota"
#> f84a5455a4f6fb3979f21796cdff4da4 "d_Bacteria" "p_Pseudomonadota"
#> c2a787cf867ebfdce216002bcc8a94e1 "d_Bacteria" "p_Pseudomonadota"
#> a0b1f105f827e856c5fdb5f98eaecc09 "d_Bacteria" "p_Pseudomonadota"
#>                                  Class                   Order              
#> 7ff346973a282aa55de296afdb5d74af "c_Actinomycetia"       "o_Actinomycetales"
#> 0f18144d308ada95632ab5193d92073f "c_Gammaproteobacteria" "o_Pseudomonadales"
#> 496ecde24f9ab698992413d3d4f04b5f "c_Gammaproteobacteria" "o_Xanthomonadales"
#> 9af3467db68cf6063627304cecd46a65 "c_Gammaproteobacteria" "o_Pseudomonadales"
#> 07da22cab5c55a4badb462bcbbded642 "c_Alphaproteobacteria" "o_Rhizobiales"    
#> 1bc599cac5647c5c30496e93ca551557 "c_Gammaproteobacteria" "o_Pseudomonadales"
#> 08bdc6c81f6b9daab3d132b54e1d45bc "c_Actinomycetia"       "o_Mycobacteriales"
#> f84a5455a4f6fb3979f21796cdff4da4 "c_Alphaproteobacteria" "o_Caulobacterales"
#> c2a787cf867ebfdce216002bcc8a94e1 "c_Alphaproteobacteria" "o_Rhizobiales"    
#> a0b1f105f827e856c5fdb5f98eaecc09 "c_Gammaproteobacteria" "o_Burkholderiales"
#>                                  Family                 Genus               
#> 7ff346973a282aa55de296afdb5d74af "f_Micrococcaceae"     "g_Arthrobacter"    
#> 0f18144d308ada95632ab5193d92073f "f_Pseudomonadaceae"   "g_Pseudomonas_E"   
#> 496ecde24f9ab698992413d3d4f04b5f "f_Xanthomonadaceae"   "g_Stenotrophomonas"
#> 9af3467db68cf6063627304cecd46a65 "f_Pseudomonadaceae"   "g_Pseudomonas_E"   
#> 07da22cab5c55a4badb462bcbbded642 "f_Rhizobiaceae"       "g_Phyllobacterium" 
#> 1bc599cac5647c5c30496e93ca551557 "f_Pseudomonadaceae"   "g_Pseudomonas_E"   
#> 08bdc6c81f6b9daab3d132b54e1d45bc "f_Mycobacteriaceae"   "g_Rhodococcus"     
#> f84a5455a4f6fb3979f21796cdff4da4 "f_Caulobacteraceae"   "g_Brevundimonas"   
#> c2a787cf867ebfdce216002bcc8a94e1 "uncl_o_Rhizobiales"   "uncl_o_Rhizobiales"
#> a0b1f105f827e856c5fdb5f98eaecc09 "f_Burkholderiaceae_B" "g_Variovorax"